requirements.txt
python-dotenv==0.20.0
twitchio==2.2.0
requests==2.27.1